# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/connections/github/start": {
      "post": {
        "description": "Génère une URL d’autorisation et un paramètre state pour connecter l’utilisateur authentifié à GitHub.",
        "operationId": "ConnectionsController_startGithub",
        "parameters": [],
        "responses": {
          "200": {
            "description": "URL d’autorisation générée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubAuthorizeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initier la connexion OAuth GitHub",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/connections/discord/start": {
      "post": {
        "description": "Génère une URL d’autorisation pour installer le bot et autoriser l’accès aux serveurs de l’utilisateur.",
        "operationId": "ConnectionsController_startDiscord",
        "parameters": [],
        "responses": {
          "200": {
            "description": "URL Discord générée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscordAuthorizeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initier la connexion OAuth Discord",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/discord/complete": {
      "post": {
        "description": "Échange le code contre un jeton utilisateur, vérifie le serveur sélectionné et enregistre la connexion.",
        "operationId": "ConnectionsController_completeDiscord",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDiscordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compte Discord connecté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscordConnectionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "State invalide ou code refusé par Discord",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finaliser la connexion Discord",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/discord/guilds": {
      "get": {
        "description": "Renvoie les guildes (serveurs) sur lesquels l'utilisateur a autorisé le bot.",
        "operationId": "ConnectionsController_listDiscordGuilds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des guildes Discord accessibles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscordGuildListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les serveurs Discord connectés",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections/discord/guilds/{guildId}/channels": {
      "get": {
        "description": "Renvoie les canaux textuels où le bot peut écrire ou réagir pour le serveur sélectionné.",
        "operationId": "ConnectionsController_listDiscordChannels",
        "parameters": [
          {
            "name": "guildId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Canaux Discord disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscordChannelListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Serveur Discord inconnu ou non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les canaux Discord d’un serveur",
        "tags": [
          "connections"
        ]
      }
    },
    "/connections": {
      "get": {
        "operationId": "ConnectionsController_listConnections",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statut de connexion actuel pour chaque fournisseur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les fournisseurs externes connectés par l’utilisateur",
        "tags": [
          "connections"
        ]
      }
    }
  },
  "info": {
    "title": "Area API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "example": [
              "le nom ne doit pas être vide",
              "l’identifiant d’action doit être un entier"
            ],
            "description": "Messages de validation détaillés",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Action invalide",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "UnauthorizedResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Non autorisé",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Identifiants invalides",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "GithubAuthorizeResponseDto": {
        "type": "object",
        "properties": {
          "authorizeUrl": {
            "type": "string",
            "example": "https://github.com/login/oauth/authorize?client_id=abc123&scope=repo%20user&state=0a7f599e-62a3-4b35-bc1c-374c4511eddf",
            "description": "URL vers laquelle rediriger l’utilisateur pour autoriser GitHub"
          },
          "state": {
            "type": "string",
            "example": "0a7f599e-62a3-4b35-bc1c-374c4511eddf",
            "description": "Valeur opaque utilisée pour valider le retour OAuth"
          }
        },
        "required": [
          "authorizeUrl",
          "state"
        ]
      },
...
      "DiscordAuthorizeResponseDto": {
        "type": "object",
        "properties": {
          "authorizeUrl": {
            "type": "string",
            "example": "https://discord.com/api/oauth2/authorize?client_id=123&scope=identify%20guilds%20bot&state=xyz",
            "description": "URL d'autorisation à ouvrir pour connecter Discord"
          },
          "state": {
            "type": "string",
            "example": "xyz-0a7f599e-62a3-4b35-bc1c-374c4511eddf",
            "description": "Valeur state utilisée pour sécuriser le callback"
          }
        },
        "required": [
          "authorizeUrl",
          "state"
        ]
      },
      "CompleteDiscordDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "1Zf9X2...",
            "description": "Code d’autorisation retourné par Discord"
          },
          "state": {
            "type": "string",
            "example": "0a7f599e-62a3-4b35-bc1c-374c4511eddf",
            "description": "Valeur state émise lors du démarrage du flux"
          },
          "guildId": {
            "type": "string",
            "example": "123456789012345678",
            "description": "Identifiant du serveur (guild) sélectionné pendant l’autorisation"
          }
        },
        "required": [
          "code",
          "state"
        ]
      },
      "DiscordAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123456789012345678"
          },
          "username": {
            "type": "string",
            "example": "JeanDupont"
          },
          "displayName": {
            "type": "object",
            "example": "Jean Dupont",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "example": "https://cdn.discordapp.com/avatars/123/abc.png",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "DiscordGuildDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "987654321098765432"
          },
          "name": {
            "type": "string",
            "example": "Serveur Communauté"
          },
          "owner": {
            "type": "boolean",
            "example": true
          },
          "permissions": {
            "type": "object",
            "example": "3072",
            "nullable": true,
            "description": "Permissions accordées au bot sous forme de bitfield décimal"
          }
        },
        "required": [
          "id",
          "name",
          "owner",
          "permissions"
        ]
      },
      "DiscordConnectionResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "provider": {
            "type": "string",
            "example": "discord"
          },
          "account": {
            "$ref": "#/components/schemas/DiscordAccountDto"
          },
          "guild": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DiscordGuildDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "provider",
          "account",
          "guild"
        ]
      },
      "DiscordGuildListResponseDto": {
        "type": "object",
        "properties": {
          "guilds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscordGuildDto"
            }
          }
        },
        "required": [
          "guilds"
        ]
      },
      "DiscordChannelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123456789012345678"
          },
          "name": {
            "type": "string",
            "example": "général"
          },
          "type": {
            "type": "number",
            "example": 0,
            "description": "Type de canal Discord (0 = texte, 5 = annonces, etc. — seuls les canaux textuels sont retournés)"
          },
          "parentId": {
            "type": "object",
            "example": "987654321012345678",
            "nullable": true,
            "description": "Identifiant de la catégorie parente, si présent"
          },
          "categoryName": {
            "type": "object",
            "example": "Catégorie Projet",
            "nullable": true,
            "description": "Nom de la catégorie qui contient ce canal"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "parentId",
          "categoryName"
        ]
      },
      "DiscordChannelListResponseDto": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscordChannelDto"
            }
          }
        },
        "required": [
          "channels"
        ]
      },
      "ConnectionStatusDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "github"
          },
          "connected": {
            "type": "boolean",
            "example": true
          },
          "connectedAt": {
            "type": "object",
            "example": "2025-01-10T12:00:00.000Z",
            "nullable": true,
            "description": "Date de connexion au fournisseur (null si la connexion n’est pas active)."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Informations supplémentaires propres au provider (identifiants, métadonnées)."
          }
        },
        "required": [
          "provider",
          "connected",
          "connectedAt"
        ]
      },
      "ConnectionsListResponseDto": {
        "type": "object",
        "properties": {
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionStatusDto"
            }
          }
        },
        "required": [
          "connections"
        ]
      }
    }
  }
}